api/python-ai/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
joblib>=1.1.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0
pymongo>=4.0.0
python-dotenv>=0.19.0
flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
requests>=2.25.0
nltk>=3.6.0
textblob>=0.15.0
wordcloud>=1.8.0
tqdm>=4.62.0
